Cleaning dependencies on Error class and renamed the base custom exception that represents business errors
