+++ application/src/main/java/com/danilat/scorecards/shared/domain/Error.java
@@ -1,18 +1,10 @@
 package com.danilat.scorecards.shared.domain;
 
-import javax.validation.ConstraintViolation;
-
 public class Error {
 
   private final String fieldName;
   private final String message;
   private final String messageTemplate;
-
-  public Error(ConstraintViolation violation) {
-    fieldName = violation.getPropertyPath().toString();
-    message = violation.getMessage();
-    messageTemplate = violation.getMessageTemplate();
-  }
 
   public Error(String fieldName, String message, String messageTemplate) {
     this.fieldName = fieldName;
