+++ application/src/main/java/com/danilat/scorecards/shared/domain/ScoreCardsException.java
@@ -1,0 +1,5 @@
+package com.danilat.scorecards.shared.domain;
+
+public abstract class ScoreCardsException extends RuntimeException {
+
+}
